show expeneses and notes to managers

+++ database/migrations/2018_05_24_192021_create_expenses_table.php
@@ -13,7 +13,7 @@ class CreateExpensesTable extends Migration
 
             $table->integer('check');
             $table->date('date');
-            $table->integer('amount');
+            $table->double('amount');
             $table->string('concept');
             $table->string('description')->nullable();
             $table->integer('store_id');
